requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.2
django-axes==6.0.3
django-environ==0.10.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
oauthlib==3.2.2
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
testfixtures==7.1.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==2.0.3
zope.interface==6.0
psycopg2_binary==2.9.6
